Updated Handicap to return Float handicaps. Used these float handicaps to better predict scores

+++ app/src/main/java/eywa/projectcodex/common/utils/updateDefaultRounds/jsonObjects/DefaultRoundInfo.kt
@@ -86,7 +86,7 @@ class DefaultRoundInfo(
      */
     class RoundInfoArrowCount(
             val distanceNumber: Int,
-            val faceSizeInCm: Double,
+            val faceSizeInCm: Float,
             val arrowCount: Int
     ) {
         fun toRoundArrowCount(roundId: Int): RoundArrowCount {
